Integer conversion done

+++ c/motor/mmtest/src/mmmotor_settings.h
@@ -13,7 +13,7 @@
 #ifndef MMMOTOR_SETTINGS_H
 #define	MMMOTOR_SETTINGS_H
 
-typedef float CHANNEL_VAL;
+typedef int16_t CHANNEL_VAL;
 
 #define CHANNELS    8      // nr of channels
 #define ALARMS      10      // nr of alarms
